backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
PyJWT==2.8.0
argon2-cffi==23.1.0
cryptography==41.0.7
requests==2.31.0
opencv-python==4.8.1.78
mediapipe==0.10.33
numpy==1.24.3
webauthn==2.0.0

# Optional: Post-Quantum Cryptography
# Requires liboqs to be installed first
# See: https://github.com/open-quantum-safe/liboqs-python
# liboqs-python==0.9.0

# Optional: Fingerprint sensor support
# pyfingerprint==1.5
# pyserial==3.5